// src/app/shared/interfaces/orders.interface.ts
export interface IOrder {
    id: number;
    name: string;
    phoneNumber: string;
    email: string;
    viber: string;
    brand: string;
    model: string;
    year: string;
    engine: string;
    fuel: string;
    probig: string;
    transmission: string;
    bodyType: string;
    price: string;
    interior: string;
    interiorColor: string;
    message: string;
}